bigquery/requirements.txt
# BigQuery dependencies for BigQuery CDC Demo
google-cloud-bigquery>=3.0.0
PyYAML>=6.0
